Add two div containers and three buttons
These buttons are the clear and delete buttons as well as the equal button. The importance
of the equals button cannot be overstated in this program, as it will be responsible for
executing several actions. This may change as development progresses, but as of now it
gets its own container.