Add memory object and modify setOperand function
Added the memory object which will store input values and operand for use in the calculation
Also modified the setOperand function to run a check if the operator value is an empty string
Future expansion will create an else clause that fires the calculate function if it is not
empty. This is in keeping with the sequenctiality request of the assignment in allowin numerical
values and operations to daisy chain and stack.