Add click variable with default state set to false
If the equals button is clicked the value becomes true. This is in order to fix
an issue with the setting of an operator post calculation. While this calculator works
perfectly fine for single operations and stacking operators, clicking the equals button
doesn't reset the operator as by functionality we want it to be stackable. But we also
may want to execute a different operation after, hence we need a way to detect and change
that.