Clean code for legibility and brevity
While a lot has been moved around and the order of things changed, the basic principles
of the calculator remain mostly the same as can be seen when looking at the changelog in code
The major differences added were a rounding function to prevent numerical overflow and the
implementation(proper) of the operate function which now takes values instead of grabbing
directly from the global object memory. Everything else is short hand or a more clean version
of an already existing function. Credit must be given to @michalosman for inspiration in the
implemenation of a reset function and boolean global variable. As well as the round function which
is borrowed directly from his solution.